Improves the documentation of to_ascii so that we know what it does.
